fixed check_initial_positions_players

+++ tests/board/check_initial_positions_players.cpp
@@ -30,8 +30,8 @@ int main() {
             int i, j;
             myfile >> i >> j;
 
-            err::check(spritePacMan[k].getPosition().x == j * MAP_WIDTH, k);
-            err::check(spritePacMan[k].getPosition().y == i * MAP_HEIGHT, k);
+            err::check((int)spritePacMan[k].getPosition().x == j * MAP_WIDTH, k);
+            err::check((int)spritePacMan[k].getPosition().y == i * MAP_HEIGHT, k);
         }
         myfile.close();
     } else {
